+++ app/Models/ChatSession.php
@@ -539,3 +539,5 @@ class ChatSession extends Model
 
 
 
+
+