+++ database/migrations/2014_10_12_100000_create_password_reset_tokens_table.php
@@ -53,3 +53,5 @@ return new class extends Migration
 
 
 
+
+