+++ database/migrations/2019_08_19_000000_create_failed_jobs_table.php
@@ -57,3 +57,5 @@ return new class extends Migration
 
 
 
+
+