+++ database/migrations/2019_12_14_000001_create_personal_access_tokens_table.php
@@ -58,3 +58,5 @@ return new class extends Migration
 
 
 
+
+