+++ database/migrations/2025_11_04_000001_create_user_profiles_table.php
@@ -89,3 +89,5 @@ return new class extends Migration
 
 
 
+
+