+++ database/migrations/2025_11_04_000004_create_user_memberships_table.php
@@ -62,3 +62,5 @@ return new class extends Migration
 
 
 
+
+