+++ database/migrations/2025_11_04_000005_create_exercises_table.php
@@ -100,3 +100,5 @@ return new class extends Migration
 
 
 
+
+