+++ database/migrations/2025_11_04_000006_create_exercise_v2_media_table.php
@@ -68,3 +68,5 @@ return new class extends Migration
 
 
 
+
+