+++ database/migrations/2025_11_04_000007_create_training_plans_table.php
@@ -73,3 +73,5 @@ return new class extends Migration
 
 
 
+
+