+++ database/migrations/2025_11_04_000009_create_training_records_table.php
@@ -70,3 +70,5 @@ return new class extends Migration
 
 
 
+
+