+++ database/migrations/2025_11_04_000010_create_training_progress_table.php
@@ -71,3 +71,5 @@ return new class extends Migration
 
 
 
+
+