+++ database/migrations/2025_11_04_000011_create_chat_sessions_table.php
@@ -80,3 +80,5 @@ return new class extends Migration
 
 
 
+
+