+++ database/seeders/UserSeeder.php
@@ -374,3 +374,5 @@ class UserSeeder extends Seeder
 
 
 
+
+